Make checking for usage of "nm_funcs" more precise.
In the push-based runtime, we need to handle rules that invoke "nm_funcs"
specially. That is, functions that might return a different value the next time
they are called (e.g., budtime). This is an abuse of the term "non-monotonic",
but that's a separate matter. This was originally implemented in commit
1a4352e501a6384318d3e1ea752893d30fe0eb65.

The previous analysis that checked for nm_funcs was too conservative, in that
any rule that did aggregation was regarded as invoking such an unsafe
function. Naturally aggregation rules are non-monotone (in the correct sense),
but they don't necessarily invoke unsafe functions.

Now that our analysis of nm_funcs is more precise, it turns out that the
optimization implemented in commit 715f86b5fc226cfd343eb11a2d3772e9446524e9 is
not correct (unless the rule happens to also invoke an unsafe function). So for
now, disable that optimization; I'll revisit this shortly.