Support code blocks in BudChannel#payloads.
Previously, a code block could be passed but was not evaluated. Along the way,
revise how we strip location specifiers from the output of the payloads method.

Note that this change interacts poorly with the attr name => number rewrite: the
rewrite mistakenly thinks it can use the attr name => number mapping for the
channel inside the code block passed to payloads(). This is obviously not right,
but fixing it properly would take a while; for now I just disabled the attr name
rewrite for the payloads code block. The proper fix is to kill the attr name
rewrite altogether, as it is no longer useful.