Initial hacky support for reducing lset copies.
In the current mainline lattice code, merge() is always a non-destructive
operation -- i.e., we create a new lattice element that contains the LUB of the
merge() inputs. That is safe (and easy to reason about), but incurs a
performance cost (unnecessary copies), especially when lattice objects are
large. This change is a quick hack to avoid redundant copies in one particular
place: when merging one lset lattice into another on the LHS of a Bloom
statement.

This is a hack, for a few different reasons:

   (1) shouldn't special-case lset

   (2) probably more circumstances in which we can avoid copies

   (3) need to think more carefully about when we can avoid copies without
       harming safety

   (4) currently, supporting "unsafe" (destructive) merges requires separate
       support in each lattice type; it would be better to have a single generic
       merge operation, and then have the runtime apply it either destructively
       or not as needed.