Fix error when output of notin sent to a lattice identifier.
The code that routed output tuples for "notin" was never updated to account for
the possibility that the output collection might be a lattice. As it turns out,
there is no reason for notin to use its own implementation of this code; simply
deleting it and relying on the parent-class push_out() implementation fixes the
problem and removes some redundant code at the same time.