Partial fix for rescan/invalidate logic + deletions.
The rescan/invalidation logic runs into several different problems when
presented with deletions to a persistent table that is downstream of other
persistent tables -- see #303. This commit fixes one of those issues: when
computing the invalidation scheme, we should consider what needs to be done when
each table is invalidated, NOT when each table is marked for rescan.

Other related issues will be addressed in subsequent commits.