Fix rescan/invalidation for tables with deletion and <+ rules.
Fix #303; see also related commits 9d8c4ab7347c5c872ed75912f7780504043ad98d and
17803ed44533578f7b56bb6e3891e0fb846b0c89. The issue addressed in this commit
involves situations in which there is a deletion on a table, and some of the
contents of the table is supplied from another table via <+ rules. We neglected
to mark such upstream elements as requiring rescan when the downstream table is
invalidated.