lset#eqjoin improvement -- make user-supplied code block optional.
It's not quite clear what we ought to do here: we naturally want to preserve
both join inputs, but we could do so by producing a pair of tuples, or producing
a single "flattened" tuple. The current behavior is to produce a pair (the user
can always flatten if desired), but that could be changed.