Micro-optimization for PushElement#push_out.
We shouldn't need to check whether the input tuple is `nil`; if it is, we
shouldn't be trying to emit it via `push_out` in the first place. Since
`push_out` is a hotspot, this discipline seems worth enforcing -- on a simple
test case, it avoided ~100k of the ~225k calls to Kernel#nil? that were made,
not to mention the conditional branch itself.