Unbreak signal-handling unit tests
The previous coding was simply wrong (unclear to me how it could have ever
passed). The current behavior seems fine, with the exception that the child
process exits with status 1, whereas it should really exit with status 0. I
poked around to see why this is happening but couldn't find an easy fix; for
now, just assert the observed (but incorrect) behavior; we can fix this later.