Extend RSE analysis slightly.
We can still apply RSE deletion to a collection when that collection appears on
the RHS of a rule whose LHS is a persistent table, BUT that persistent table has
a deletion rule, BUT that deletion rule was created by RSE in the first
place. That is, we require the downstream consumers of the reclaimed relation to
be conceptually inflationary; RSE-induced deletion rules should not harm this
property.