Extend notin qual syntax.
Allow notin quals to be specified via position: that is,

    foo.notin(bar, 3 => :x)

Specifies that the column at offset 3 in `foo` should be matched against the
column named `x` in `bar`. This is useful when the left input to the notin does
not have a well-defined schema (e.g., when it is a join).