Optimize RCE to only send key values in ACKs.
This is sufficient, because the keys (at the sender) functionally determine the
rest of the outbound message. By only sending keys, we reduce the size of the
ACKs; moreover, this means the sender-side negation need consider fewer columns,
which will allow us to implement it more efficiently in the future.