Improvements for RSE.
Given

    x <= a.notin(b)
    y <= a.notin(c)

When y is not persistent, we would normally treat it is an output relation,
which would mean we couldn't reclaim inputs that we need to reproduce
y. However, the rule that computes y is a negation; hence we can reclaim x
tuples when a matching tuple appears in both b and c, regardless of whether y is
persistent.