Fix a bug in chaining multiple notins after a join.
Somewhat obscure, but arises in RSE/GC-generated code. Basic problem is that we
only want to call setup_accessors once; for lefts/rights, we were calling it
multiple times. Unclear why that mucked things up, but it was certainly wrong
anyway.

On examination, the code for assigning a schema to a pairs join is completely
bogus anyway (see comment) -- but I won't fix that for now.