Refactor RSE code.
Introduce the concept of an RseChangeManager, which does the heavy lifting
involved in actually installing rule and collection definitions into the Bud
runtime. Right now the RseChangeManager does essentially nothing; but in the
future we'll refactor it so that (a) new definitions aren't introduced until the
end of the rewrite process (b) redundant definitions are avoided.