Improve RSE rule optimizer.
We now support a new class of rewrites: if the site of *usage* for a relation is
simple, we can replace the rel with its definition, even if the definition
itself is not simple. That is, we now optimize relations defined by a single
rule, where *either* the definition of the rel or the usage of the rel is
simple.