Implement a new RSE analysis, motivated by the causal KVS.
In a join like:

    x <= (t1 * t2).pairs(:k1 => :k2) { ... }

Where "k2" is the entire key of t2, we effectively have a seal on t2 -- we know
that no other t2 tuple will appear that can join with a t1 tuple that already
has a match in t2. Hence, we can use this information to do more effective
sealing.