Quick hack to get GC for causal KVS fixed.
Basic problem is that when there's a join dependency for X on Y that we satisfy
via a key condition, we check for the presence of a tuple in Y matching the
key. That works fine, except when the matching Y tuple is subsequently reclaimed
and THEN another copy of the X tuple is delivered.

Short-term fix is to check for Y_keys; if it was already created for another
purpose, we can use it. Probably the right fix is to infer a seal tuple when the
Y tuple is observed, rather than creating a separate reclamation rule on Y.